Split tests into designated files
